+++ src/components/base/b-router/b-router.ts
@@ -424,8 +424,7 @@ export default class bRouter extends iRouterProps {
 	protected override initBaseAPI(): void {
 		super.initBaseAPI();
 
-		const
-			i = this.instance;
+		const i = this.instance;
 
 		this.compileStaticRoutes = i.compileStaticRoutes.bind(this);
 		this.emitTransition = i.emitTransition.bind(this);