+++ src/components/super/i-block/event/interface.ts
@@ -9,27 +9,29 @@
 import type { LogLevel } from 'core/log';
 
 export type InferEvents<
-	I extends Array<[string, ...any[]]>,
-	P extends Dictionary = {},
-	R extends Dictionary = {}
+	Scheme extends Array<[string, ...any[]]>,
+	Parent extends Dictionary = {}
+> = _InferEvents<FlatEvents<Scheme>, Parent>;
+
+export type _InferEvents<
+	Scheme extends any[],
+	Parent extends Dictionary = {},
+	Result extends Dictionary = {}
 > = {
-	0: InferEvents<TB.Tail<I>, P, (TB.Head<I> extends [infer E, ...infer A] ?
-		E extends string ? {
-			Args: {[K in E]: A};
+	0: _InferEvents<TB.Tail<Scheme>, Parent, (TB.Head<Scheme> extends [infer E, ...infer A] ? E extends string ? {
+		on(event: E, cb: (...args: A) => void): void;
+		once(event: E, cb: (...args: A) => void): void;
+		promisifyOnce(event: E): Promise<CanUndef<TB.Head<A>>>;
 
-			on(event: E, cb: (...args: A) => void): void;
-			once(event: E, cb: (...args: A) => void): void;
-			promisifyOnce(event: E): Promise<CanUndef<TB.Head<A>>>;
+		off(event: E | string, handler?: Function): void;
 
-			off(event: E | string, handler?: Function): void;
+		strictEmit(event: E, ...args: A): void;
+		emit(event: E, ...args: A): void;
+		emit(event: string, ...args: unknown[]): void;
+	} : {} : {}) & Result>;
 
-			strictEmit(event: E, ...args: A): void;
-			emit(event: E, ...args: A): void;
-			emit(event: string, ...args: unknown[]): void;
-		} : {} : {}) & R>;
-
-	1: R & P;
-}[TB.Length<I> extends 0 ? 1 : 0];
+	1: Result & Parent;
+}[TB.Length<Scheme> extends 0 ? 1 : 0];
 
 export interface ComponentEvent<E extends string = string> {
 	event: E;
@@ -37,50 +39,51 @@ export interface ComponentEvent<E extends string = string> {
 }
 
 export type InferComponentEvents<
-	C,
-	I extends Array<[string, ...any[]]>,
-	P extends Dictionary = {},
-	R extends Dictionary = {}
+	Scheme extends Array<[string, ...any[]]>,
+	Parent extends Dictionary = {}
+> = _InferComponentEvents<FlatEvents<Scheme>, Parent>;
+
+export type _InferComponentEvents<
+	Scheme extends any[],
+	Parent extends Dictionary = {},
+	Result extends Dictionary = {}
 > = {
-	0: InferComponentEvents<C, TB.Tail<I>, P, (TB.Head<I> extends [infer E, ...infer A] ? E extends string ? {
-		Args: {[K in E]: A};
+	0: _InferComponentEvents<
+		TB.Tail<Scheme>,
 
-		on(event: `on${Capitalize<E>}`, cb: (...args: A) => void): void;
-		on(event: E | `${E}:component`, cb: (component: C, ...args: A) => void): void;
+		Parent,
 
-		once(event: `on${Capitalize<E>}`, cb: (...args: A) => void): void;
-		once(event: E | `${E}:component`, cb: (component: C, ...args: A) => void): void;
+		(TB.Head<Scheme> extends [infer E, ...infer A] ? E extends string ? {
+			on(event: `on${Capitalize<E>}`, cb: (...args: A) => void): void;
+			on<T>(this: T, event: E | `${E}:component`, cb: (component: T, ...args: A) => void): void;
 
-		promisifyOnce(event: `on${Capitalize<E>}`): Promise<CanUndef<TB.Head<A>>>;
-		promisifyOnce(event: E | `${E}:component`): Promise<CanUndef<C>>;
+			once(event: `on${Capitalize<E>}`, cb: (...args: A) => void): void;
+			once<T>(this: T, event: E | `${E}:component`, cb: (component: T, ...args: A) => void): void;
 
-		off(event: E | `${E}:component` | `on${Capitalize<E>}` | string, handler?: Function): void;
+			promisifyOnce(event: `on${Capitalize<E>}`): Promise<CanUndef<TB.Head<A>>>;
+			promisifyOnce<T>(this: T, event: E | `${E}:component`): Promise<CanUndef<T>>;
 
-		strictEmit(event: E | ComponentEvent<E>, ...args: A): void;
-		emit(event: E | ComponentEvent<E>, ...args: A): void;
-		emit(event: string | ComponentEvent, ...args: unknown[]): void;
-	} : {} : {}) & R>;
+			off(event: E | `${E}:component` | `on${Capitalize<E>}` | string, handler?: Function): void;
 
-	1: R & OverrideParentComponentEvents<C, P>;
-}[TB.Length<I> extends 0 ? 1 : 0];
+			strictEmit(event: E | ComponentEvent<E>, ...args: A): void;
+			emit(event: E | ComponentEvent<E>, ...args: A): void;
+			emit(event: string | ComponentEvent, ...args: unknown[]): void;
+		} : {} : {}) & Result>;
 
-export type OverrideParentComponentEvents<C, P extends Dictionary, A = P['Args']> = A extends Record<string, any[]> ? {
-	[E in keyof A]: E extends string ? {
-		Args: A;
+	1: Parent & Result;
+}[TB.Length<Scheme> extends 0 ? 1 : 0];
 
-		on(event: `on${Capitalize<E>}`, cb: (...args: A[E]) => void): void;
-		on(event: E | `${E}:component`, cb: (component: C, ...args: A[E]) => void): void;
+type UnionToIntersection<T> = (T extends any ? (k: T) => void : never) extends (k: infer R) => void ? R : never;
 
-		once(event: `on${Capitalize<E>}`, cb: (...args: A[E]) => void): void;
-		once(event: E | `${E}:component`, cb: (component: C, ...args: A[E]) => void): void;
+type EventToTuple<Event, Result extends any[] = []> =
+	UnionToIntersection<Event extends any ? () => Event : never> extends (() => infer T) ?
+		[...EventToTuple<Exclude<Event, T>, Result>, [T, ...Result]] :
+		[];
 
-		promisifyOnce(event: `on${Capitalize<E>}`): Promise<CanUndef<TB.Head<A[E]>>>;
-		promisifyOnce(event: E | `${E}:component`): Promise<CanUndef<C>>;
+export type FlatEvents<Events extends any[], Result extends any[] = []> = {
+	0: TB.Head<Events> extends [infer E, ...infer A] ?
+		FlatEvents<TB.Tail<Events>, [...EventToTuple<E, A>, ...Result]> :
+		[];
 
-		off(event: E | `${E}:component` | `on${Capitalize<E>}` | string, handler?: Function): void;
-
-		strictEmit(event: E | ComponentEvent<E>, ...args: A[E]): void;
-		emit(event: E | ComponentEvent<E>, ...args: A[E]): void;
-		emit(event: string | ComponentEvent, ...args: unknown[]): void;
-	} : {};
-}[keyof A] : {};
+	1: Result;
+}[TB.Length<Events> extends 0 ? 1 : 0];