+++ src/core/component/directives/attrs/index.ts
@@ -65,8 +65,7 @@ ComponentEngine.directive('attrs', {
 			r = ctx.$renderEngine.r;
 		}
 
-		let
-			attrs = {...params.value};
+		let attrs = {...params.value};
 
 		if (componentMeta != null) {
 			attrs = normalizeComponentAttrs(attrs, vnode.dynamicProps, componentMeta)!;
@@ -113,21 +112,17 @@ ComponentEngine.directive('attrs', {
 		}
 
 		function parseDirective(attrName: string, attrVal: unknown) {
-			const
-				decl = directiveRgxp.exec(attrName);
+			const decl = directiveRgxp.exec(attrName);
 
-			let
-				value = attrVal;
+			let value = attrVal;
 
 			if (decl == null) {
 				throw new SyntaxError('Invalid directive declaration');
 			}
 
-			const
-				[, name, arg = '', rawModifiers = ''] = decl;
+			const [, name, arg = '', rawModifiers = ''] = decl;
 
-			let
-				dir: CanUndef<object>;
+			let dir: CanUndef<object>;
 
 			switch (name) {
 				case 'show': {
@@ -137,9 +132,9 @@ ComponentEngine.directive('attrs', {
 
 				case 'on': {
 					if (Object.isDictionary(value)) {
-						Object.entries(value).forEach(([name, handler]) => {
-							attachEvent(name, handler);
-						});
+						for (const name of Object.keys(value)) {
+							attachEvent(name, value[name]);
+						}
 					}
 
 					return;
@@ -147,10 +142,10 @@ ComponentEngine.directive('attrs', {
 
 				case 'bind': {
 					if (Object.isDictionary(value)) {
-						Object.entries(value).forEach(([name, val]) => {
-							attrs[name] = val;
+						for (const name of Object.keys(value)) {
+							attrs[name] = value[name];
 							attrsKeys.push(name);
-						});
+						}
 					}
 
 					return;
@@ -165,8 +160,7 @@ ComponentEngine.directive('attrs', {
 						handlerCache = getHandlerStore(),
 						handlerKey = `onUpdate:${modelProp}:${modelValLink}`;
 
-					let
-						handler = handlerCache.get(handlerKey);
+					let handler = handlerCache.get(handlerKey);
 
 					if (handler == null) {
 						handler = (newVal: unknown) => {
@@ -231,8 +225,7 @@ ComponentEngine.directive('attrs', {
 
 			if (Object.isDictionary(dir)) {
 				if (Object.isFunction(dir.beforeCreate)) {
-					const
-						newVnode = dir.beforeCreate(binding, vnode);
+					const newVnode = dir.beforeCreate(binding, vnode);
 
 					if (newVnode != null) {
 						vnode = newVnode;
@@ -287,8 +280,7 @@ ComponentEngine.directive('attrs', {
 			props: Dictionary = vnode?.props ?? {},
 			componentMeta = ctx?.meta;
 
-		let
-			attrs = {...params.value};
+		let attrs = {...params.value};
 
 		if (vnode != null) {
 			vnode.props ??= props;
@@ -298,30 +290,29 @@ ComponentEngine.directive('attrs', {
 			attrs = normalizeComponentAttrs(attrs, null, componentMeta)!;
 		}
 
-		Object.entries(attrs).forEach(([name, value]) => {
+		for (const name of Object.keys(attrs)) {
+			const value = attrs[name];
+
 			if (name.startsWith('v-')) {
 				parseDirective(name, value);
 
 			} else if (!name.startsWith('@') || isPropGetter.test(name)) {
 				patchProps(props, normalizePropertyAttribute(name), value);
 			}
-		});
+		}
 
 		return props;
 
 		function parseDirective(attrName: string, attrVal: unknown) {
-			const
-				decl = directiveRgxp.exec(attrName);
+			const decl = directiveRgxp.exec(attrName);
 
 			if (decl == null) {
 				throw new SyntaxError('Invalid directive declaration');
 			}
 
-			const
-				[, name, arg = '', rawModifiers = ''] = decl;
+			const [, name, arg = '', rawModifiers = ''] = decl;
 
-			let
-				dir: CanUndef<object>;
+			let dir: CanUndef<object>;
 
 			switch (name) {
 				case 'show': {
@@ -331,9 +322,9 @@ ComponentEngine.directive('attrs', {
 
 				case 'bind': {
 					if (Object.isDictionary(attrVal)) {
-						Object.entries(attrVal).forEach(([name, val]) => {
-							props[name] = val;
-						});
+						for (const name of Object.keys(attrVal)) {
+							props[name] = attrVal[name];
+						}
 					}
 
 					return;