+++ src/core/component/init/states/before-create.ts
@@ -15,7 +15,6 @@ import { V4_COMPONENT } from 'core/component/const';
 import { getComponentContext } from 'core/component/context';
 
 import { forkMeta } from 'core/component/meta';
-import { getPropertyInfo, PropertyInfo } from 'core/component/reflect';
 import { getNormalParent } from 'core/component/traverse';
 
 import { initProps, attachAttrPropsListeners } from 'core/component/prop';
@@ -51,8 +50,6 @@ export function beforeCreateState(
 ): void {
 	meta = forkMeta(meta);
 
-	const isFunctional = meta.params.functional === true;
-
 	// To avoid TS errors marks all properties as editable
 	const unsafe = Object.cast<Writable<ComponentInterface['unsafe']>>(component);
 	unsafe[V4_COMPONENT] = true;
@@ -66,6 +63,12 @@ export function beforeCreateState(
 		configurable: true,
 		enumerable: true,
 		get: () => meta.instance
+	});
+
+	Object.defineProperty(unsafe, 'constructor', {
+		configurable: true,
+		enumerable: true,
+		value: meta.constructor
 	});
 
 	unsafe.$fields = {};
@@ -116,7 +119,9 @@ export function beforeCreateState(
 				return ctx[$getRoot];
 			}
 
-			const fn = () => ('getRoot' in ctx ? ctx.getRoot?.() : null) ?? ctx.$root;
+			let fn = () => ('getRoot' in ctx ? ctx.getRoot?.() : null) ?? ctx.$root;
+
+			fn = fn.once();
 
 			Object.defineProperty(ctx, $getRoot, {
 				configurable: true,
@@ -129,14 +134,18 @@ export function beforeCreateState(
 		})
 	});
 
+	let r: CanNull<ComponentInterface['Root']> = null;
+
 	Object.defineProperty(unsafe, 'r', {
 		configurable: true,
 		enumerable: true,
 		get: () => {
-			const r = ('getRoot' in unsafe ? unsafe.getRoot?.() : null) ?? unsafe.$root;
+			if (r == null) {
+				r = ('getRoot' in unsafe ? unsafe.getRoot?.() : null) ?? unsafe.$root;
 
-			if ('$remoteParent' in r.unsafe) {
-				return r.unsafe.$remoteParent!.$root;
+				if ('$remoteParent' in r.unsafe) {
+					r = r.unsafe.$remoteParent!.$root;
+				}
 			}
 
 			return r;
@@ -169,6 +178,8 @@ export function beforeCreateState(
 			if (fn == null) {
 				fn = () => ctx;
 			}
+
+			fn = fn.once();
 
 			Object.defineProperty(targetCtx, $getParent, {
 				configurable: true,
@@ -208,7 +219,7 @@ export function beforeCreateState(
 
 	unsafe.$normalParent = getNormalParent(component);
 
-	['$root', '$parent', '$normalParent'].forEach((key) => {
+	for (const key of ['$root', '$parent', '$normalParent']) {
 		const val = unsafe[key];
 
 		if (val != null) {
@@ -219,7 +230,7 @@ export function beforeCreateState(
 				value: getComponentContext(Object.cast(val))
 			});
 		}
-	});
+	}
 
 	Object.defineProperty(unsafe, '$children', {
 		configurable: true,
@@ -252,9 +263,7 @@ export function beforeCreateState(
 		}());
 	});
 
-	if (opts?.addMethods) {
-		attachMethodsFromMeta(component);
-	}
+	attachMethodsFromMeta(component);
 
 	if (opts?.implementEventAPI) {
 		implementEventEmitterAPI(component);
@@ -288,86 +297,7 @@ export function beforeCreateState(
 
 	runHook('beforeRuntime', component).catch(stderr);
 
-	const {
-		systemFields,
-		tiedFields,
-
-		computedFields,
-		accessors,
-
-		watchDependencies,
-		watchers
-	} = meta;
-
-	initFields(systemFields, component, unsafe);
-
-	const fakeHandler = () => undefined;
-
-	if (watchDependencies.size > 0) {
-		const watchSet = new Set<PropertyInfo>();
-
-		watchDependencies.forEach((deps) => {
-			deps.forEach((dep) => {
-				const
-					path = Object.isArray(dep) ? dep.join('.') : String(dep),
-					info = getPropertyInfo(path, component);
-
-				if (info.type === 'system' || isFunctional && info.type === 'field') {
-					watchSet.add(info);
-				}
-			});
-		});
-
-		// If a computed property has a field or system field as a dependency
-		// and the host component does not have any watchers to this field,
-		// we need to register a "fake" watcher to enforce watching
-		watchSet.forEach((info) => {
-			const needToForceWatching =
-				watchers[info.name] == null &&
-				watchers[info.originalPath] == null &&
-				watchers[info.path] == null;
-
-			if (needToForceWatching) {
-				watchers[info.name] = [
-					{
-						deep: true,
-						immediate: true,
-						provideArgs: false,
-						handler: fakeHandler
-					}
-				];
-			}
-		});
-	}
-
-	// If a computed property is tied with a field or system field
-	// and the host component does not have any watchers to this field,
-	// we need to register a "fake" watcher to enforce watching
-	Object.entries(tiedFields).forEach(([name, normalizedName]) => {
-		if (normalizedName == null) {
-			return;
-		}
-
-		const
-			accessor = accessors[normalizedName],
-			computed = accessor == null ? computedFields[normalizedName] : null;
-
-		const needForceWatch = watchers[name] == null && (
-			accessor != null && accessor.dependencies?.length !== 0 ||
-			computed != null && computed.cache !== 'forever' && computed.dependencies?.length !== 0
-		);
-
-		if (needForceWatch) {
-			watchers[name] = [
-				{
-					deep: true,
-					immediate: true,
-					provideArgs: false,
-					handler: fakeHandler
-				}
-			];
-		}
-	});
+	initFields(meta.systemFieldInitializers, component, unsafe);
 
 	runHook('beforeCreate', component).catch(stderr);
 	callMethodFromComponent(component, 'beforeCreate');