+++ src/core/component/interface/component/component.ts
@@ -79,6 +79,12 @@ export abstract class ComponentInterface {
 	abstract readonly globalName?: string;
 
 	/**
+	 * If set to true, the component will inherit modifiers from the parent `sharedMods` property.
+	 * This prop is set automatically during the build.
+	 */
+	abstract readonly inheritMods?: boolean;
+
+	/**
 	 * True if the component renders as a regular one, but can be rendered as a functional.
 	 * This parameter is used during SSR and when hydrating the page.
 	 */
@@ -163,9 +169,9 @@ export abstract class ComponentInterface {
 	abstract readonly getParent?: () => this['$parent'];
 
 	/**
-	 * The getter is used to get a set of props that were passed to the component directly through the template
+	 * The getter is used to get a dictionary of props that were passed to the component directly through the template
 	 */
-	abstract readonly getPassedProps?: () => Set<string>;
+	abstract readonly getPassedProps?: () => Dictionary;
 
 	/**
 	 * A string value indicating the lifecycle hook that the component is currently in.
@@ -251,7 +257,7 @@ export abstract class ComponentInterface {
 	/**
 	 * A dictionary containing component attributes that are not identified as input properties
 	 */
-	protected readonly $attrs!: Dictionary<string>;
+	protected readonly $attrs!: Dictionary<string | Function>;
 
 	/**
 	 * A dictionary containing the watchable component fields that can trigger a re-rendering of the component