+++ src/core/component/render/helpers/normalizers.ts
@@ -23,20 +23,26 @@ export function normalizeClass(classes: CanArray<string | Dictionary>): string {
 		classesStr = classes;
 
 	} else if (Object.isArray(classes)) {
-		classes.forEach((className) => {
-			const normalizedClass = normalizeClass(className);
+		for (let i = 0; i < classes.length; i += 1) {
+			const
+				className = classes[i],
+				normalizedClass = normalizeClass(className);
 
 			if (normalizedClass !== '') {
 				classesStr += `${normalizedClass} `;
 			}
-		});
+		}
 
 	} else if (Object.isDictionary(classes)) {
-		Object.entries(classes).forEach(([className, has]) => {
-			if (Object.isTruly(has)) {
+		const keys = Object.keys(classes);
+
+		for (let i = 0; i < keys.length; i++) {
+			const className = keys[i];
+
+			if (Object.isTruly(classes[className])) {
 				classesStr += `${className} `;
 			}
-		});
+		}
 	}
 
 	return classesStr.trim();
@@ -50,15 +56,22 @@ export function normalizeStyle(styles: CanArray<string | Dictionary<string>>): s
 	if (Object.isArray(styles)) {
 		const normalizedStyles = {};
 
-		styles.forEach((style) => {
+		for (let i = 0; i < styles.length; i++) {
+			const style = styles[i];
+
 			const normalizedStyle = Object.isString(style) ?
 				parseStringStyle(style) :
 				normalizeStyle(style);
 
-			if (Object.size(normalizedStyle) > 0) {
-				Object.entries(normalizedStyle).forEach(([name, style]) => normalizedStyles[name] = style);
+			if (Object.isDictionary(normalizedStyle)) {
+				const keys = Object.keys(normalizedStyle);
+
+				for (let i = 0; i < keys.length; i++) {
+					const name = keys[i];
+					normalizedStyles[name] = normalizedStyle[name];
+				}
 			}
-		});
+		}
 
 		return normalizedStyles;
 	}
@@ -85,17 +98,19 @@ const
 export function parseStringStyle(style: string): Dictionary<string> {
 	const styles = {};
 
-	style.split(listDelimiterRgxp).forEach((singleStyle) => {
-		singleStyle = singleStyle.trim();
+	const styleRules = style.split(listDelimiterRgxp);
 
-		if (singleStyle !== '') {
-			const chunks = singleStyle.split(propertyDelimiterRgxp, 2);
+	for (let i = 0; i < styleRules.length; i++) {
+		const style = styleRules[i].trim();
+
+		if (style !== '') {
+			const chunks = style.split(propertyDelimiterRgxp, 2);
 
 			if (chunks.length > 1) {
 				styles[chunks[0].trim()] = chunks[1].trim();
 			}
 		}
-	});
+	}
 
 	return styles;
 }
@@ -130,12 +145,18 @@ export function normalizeComponentAttrs(
 		normalizedAttrs['v-attrs'] = normalizeComponentAttrs(normalizedAttrs['v-attrs'], dynamicProps, component);
 	}
 
-	Object.entries(normalizedAttrs).forEach(normalizeAttr);
+	const attrNames = Object.keys(normalizedAttrs);
+
+	for (let i = 0; i < attrNames.length; i++) {
+		const attrName = attrNames[i];
+		normalizeAttr(attrName, normalizedAttrs[attrName]);
+	}
+
 	modifyDynamicPath();
 
 	return normalizedAttrs;
 
-	function normalizeAttr([attrName, value]: [string, unknown]) {
+	function normalizeAttr(attrName: string, value: unknown) {
 		let propName = `${attrName}Prop`.camelize(false);
 
 		if (attrName === 'ref' || attrName === 'ref_for') {
@@ -165,7 +186,7 @@ export function normalizeComponentAttrs(
 				tiedPropValue = value()[0];
 
 			normalizedAttrs[tiedPropName] = tiedPropValue;
-			normalizeAttr([tiedPropName, tiedPropValue]);
+			normalizeAttr(tiedPropName, tiedPropValue);
 			dynamicProps.push(tiedPropName);
 
 		} else if (isGetter) {
@@ -207,8 +228,7 @@ export function normalizeComponentAttrs(
 			return;
 		}
 
-		// eslint-disable-next-line vars-on-top, no-var
-		for (var i = dynamicProps.length - 1; i >= 0; i--) {
+		for (let i = dynamicProps.length - 1; i >= 0; i--) {
 			const
 				prop = dynamicProps[i],
 				path = dynamicPropsPatches.get(prop);
@@ -228,9 +248,8 @@ export function normalizeComponentAttrs(
 }
 
 /**
- * Normalizes the props with `forceUpdate` set to `false` for a child component
- * using the parent context. The function returns a new object of normalized props
- * for the child component.
+ * Normalizes the props with forceUpdate set to false for a child component using the parent context.
+ * The function returns a new object containing the normalized props for the child component.
  *
  * @param parentCtx - the context of the parent component
  * @param componentName - the name of the child component
@@ -247,18 +266,23 @@ export function normalizeComponentForceUpdateProps(
 		return props;
 	}
 
-	const normalizedProps = {};
+	const
+		normalizedProps = {},
+		propNames = Object.keys(props);
 
-	Object.entries(props).forEach(([key, value]) => {
-		const propInfo = meta.props[key] ?? meta.props[`${key}Prop`];
+	for (let i = 0; i < propNames.length; i++) {
+		const
+			propName = propNames[i],
+			propVal = props[propName],
+			propInfo = meta.props[propName] ?? meta.props[`${propName}Prop`];
 
 		if (propInfo?.forceUpdate === false) {
-			normalizedProps[`@:${key}`] = parentCtx.unsafe.createPropAccessors(() => <object>value);
+			normalizedProps[`@:${propName}`] = parentCtx.unsafe.createPropAccessors(() => <object>propVal);
 
 		} else {
-			normalizedProps[key] = value;
+			normalizedProps[propName] = propVal;
 		}
-	});
+	}
 
 	return normalizedProps;
 }