+++ src/core/component/render/wrappers.ts
@@ -140,8 +140,8 @@ export function wrapCreateBlock<T extends typeof createBlock>(original: T): T {
 			vnode.virtualParent.value :
 			this;
 
-		let passedProps: CanNull<Set<string>> = null;
-		props.getPassedProps ??= () => passedProps ??= new Set(attrs != null ? Object.keys(attrs) : []);
+		let passedProps: Nullable<Dictionary> = null;
+		props.getPassedProps ??= () => passedProps ??= attrs;
 
 		// For refs within functional components,
 		// it is necessary to explicitly set a reference to the instance of the component
@@ -163,15 +163,19 @@ export function wrapCreateBlock<T extends typeof createBlock>(original: T): T {
 		const virtualCtx = createVirtualContext(component, {parent: this, props: attrs, slots});
 		vnode.virtualComponent = virtualCtx;
 
+		const filteredAttrs = {};
+
 		const
 			declaredProps = component.props,
-			filteredAttrs = {};
+			propKeys = Object.keys(props);
 
-		Object.entries(props).forEach(([key, val]) => {
-			if (declaredProps[key.camelize(false)] == null) {
-				filteredAttrs[key] = val;
+		for (let i = 0; i < propKeys.length; i++) {
+			const propName = propKeys[i];
+
+			if (declaredProps[propName.camelize(false)] == null) {
+				filteredAttrs[propName] = props[propName];
 			}
-		});
+		}
 
 		const functionalVNode = virtualCtx.render(virtualCtx, []);
 
@@ -208,15 +212,19 @@ export function wrapCreateBlock<T extends typeof createBlock>(original: T): T {
 		}
 
 		if (!SSR && functionalVNode.dynamicProps != null && functionalVNode.dynamicProps.length > 0) {
+			const functionalProps = functionalVNode.dynamicProps;
+
 			const dynamicProps = vnode.dynamicProps ?? [];
 			vnode.dynamicProps = dynamicProps;
 
-			functionalVNode.dynamicProps.forEach((propName) => {
+			for (let i = 0; i < functionalProps.length; i++) {
+				const propName = functionalProps[i];
+
 				if (isHandler.test(propName)) {
 					dynamicProps.push(propName);
 					setVNodePatchFlags(vnode, 'props');
 				}
-			});
+			}
 		}
 
 		functionalVNode.ignore = true;
@@ -398,7 +406,9 @@ export function wrapWithDirectives<T extends typeof withDirectives>(_: T): T {
 			Object.cast(this.$normalParent) :
 			this;
 
-		dirs.forEach((decl) => {
+		for (let i = 0; i < dirs.length; i++) {
+			const decl = dirs[i];
+
 			const [dir, value, arg, modifiers] = decl;
 
 			const binding: DirectiveBinding = {
@@ -416,7 +426,8 @@ export function wrapWithDirectives<T extends typeof withDirectives>(_: T): T {
 			};
 
 			if (!Object.isDictionary(dir)) {
-				return bindings.push(binding);
+				bindings.push(binding);
+				continue;
 			}
 
 			if (Object.isFunction(dir.beforeCreate)) {
@@ -434,7 +445,7 @@ export function wrapWithDirectives<T extends typeof withDirectives>(_: T): T {
 			} else if (Object.keys(dir).length > 0) {
 				bindings.push(binding);
 			}
-		});
+		}
 
 		return vnode;
 
@@ -536,7 +547,9 @@ export function wrapAPI<T extends Dictionary>(this: ComponentInterface, path: st
 	async function unrollBuffer(buf: BufItems): Promise<string> {
 		let res = '';
 
-		for (let val of buf) {
+		for (let i = 0; i < buf.length; i++) {
+			let val = buf[i];
+
 			if (Object.isPromise(val)) {
 				val = await val;
 			}