+++ src/core/component/watch/bind.ts
@@ -37,97 +37,97 @@ export function bindRemoteWatchers(component: ComponentInterface, params?: BindR
 		{unsafe} = component,
 		{$watch, meta, hook, meta: {hooks}} = unsafe;
 
+	const $a = p.async ?? unsafe.$async;
+
 	const
-		componentAsync = unsafe.$async,
-		$a = p.async ?? componentAsync;
+		allWatchers = p.watchers ?? meta.watchers,
+		watcherKeys = Object.keys(allWatchers);
 
 	const
 		// True if the component is currently deactivated
 		isDeactivated = hook === 'deactivated',
 
 		// True if the component has not been created yet
-		isBeforeCreate = Boolean(beforeHooks[hook]);
-
-	const
-		watchersMap = p.watchers ?? meta.watchers,
+		isBeforeCreate = Boolean(beforeHooks[hook]),
 
 		// True if the method has been invoked with passing the custom async instance as a property
-		customAsync = $a !== unsafe.$async;
+		isCustomAsync = $a !== unsafe.$async;
 
 	// Iterate over all registered watchers and listeners and initialize their
-	Object.entries(watchersMap).forEach(([watchPath, watchers]) => {
+	for (let i = 0; i < watcherKeys.length; i++) {
+		let watchPath = watcherKeys[i];
+
+		const watchers = allWatchers[watchPath];
+
 		if (watchers == null) {
-			return;
+			continue;
 		}
 
-		let
-			// A link to the context of the watcher;
-			// by default, a component is passed to the function
-			watcherCtx = component,
+		// A link to the context of the watcher;
+		// by default, a component is passed to the function
+		let watcherCtx = component;
 
-			// True if this watcher can initialize only when the component is created
-			watcherNeedCreated = true,
+		// True if this watcher can initialize only when the component is created
+		let attachWatcherOnCreated = true;
 
-			// True if this watcher can initialize only when the component is mounted
-			watcherNeedMounted = false;
+		// True if this watcher can initialize only when the component is mounted
+		let attachWatcherOnMounted = false;
 
 		// Custom watchers look like ':foo', 'bla:foo', '?bla:foo'
-		// and are used to listen to custom events instead of property mutations.
+		// and are used to listen to custom events instead of property mutations
 		const customWatcher = isCustomWatcher.test(watchPath) ? customWatcherRgxp.exec(watchPath) : null;
 
 		if (customWatcher != null) {
-			const m = customWatcher[1];
-			watcherNeedCreated = m === '';
-			watcherNeedMounted = m === '?';
+			const hookMod = customWatcher[1];
+			attachWatcherOnCreated = hookMod === '';
+			attachWatcherOnMounted = hookMod === '?';
 		}
 
-		// Add a listener to a component's created hook if the component has not yet been created
-		if (watcherNeedCreated && isBeforeCreate) {
-			hooks['before:created'].push({fn: attachWatcher});
-			return;
-		}
-
-		// Add a listener to a component's mounted/activated hook if the component has not yet been mounted or activated
-		if (watcherNeedMounted && (isBeforeCreate || component.$el == null)) {
-			hooks[isDeactivated ? 'activated' : 'mounted'].unshift({fn: attachWatcher});
-			return;
-		}
-
-		attachWatcher();
-
-		function attachWatcher() {
+		const attachWatcher = () => {
 			// If we have a custom watcher, we need to find a link to the event emitter.
 			// For instance:
 			// ':foo' -> watcherCtx == ctx; key = 'foo'
 			// 'document:foo' -> watcherCtx == document; key = 'foo'
 			if (customWatcher != null) {
-				const l = customWatcher[2];
+				const pathToEmitter = customWatcher[2];
 
-				if (l !== '') {
-					watcherCtx = Object.get(component, l) ?? Object.get(globalThis, l) ?? component;
-					watchPath = customWatcher[3].toString();
+				if (pathToEmitter !== '') {
+					watcherCtx = Object.get(component, pathToEmitter) ?? Object.get(globalThis, pathToEmitter) ?? component;
 
 				} else {
 					watcherCtx = component;
-					watchPath = customWatcher[3].dasherize();
 				}
+
+				watchPath = customWatcher[3];
 			}
 
+			let propInfo: typeof p.info = p.info;
+
 			// Iterates over all registered handlers for this watcher
-			watchers!.forEach((watchInfo) => {
+			for (let i = 0; i < watchers.length; i++) {
+				const watchInfo = watchers[i];
+
 				if (watchInfo.shouldInit?.(component) === false) {
-					return;
+					continue;
+				}
+
+				if (customWatcher == null) {
+					propInfo ??= getPropertyInfo(watchPath, component);
+
+					if (canSkipWatching(propInfo, watchInfo)) {
+						continue;
+					}
 				}
 
 				const rawHandler = watchInfo.handler;
 
 				const asyncParams = {
-					label: watchInfo.label,
 					group: watchInfo.group,
+					label: watchInfo.label,
 					join: watchInfo.join
 				};
 
-				if (!customAsync) {
+				if (!isCustomAsync) {
 					if (asyncParams.label == null && !watchInfo.immediate) {
 						let defLabel: string;
 
@@ -257,36 +257,28 @@ export function bindRemoteWatchers(component: ComponentInterface, params?: BindR
 						}
 
 						if (customWatcher) {
-							// True if the component itself can listen to an event,
-							// because watcherCtx does not appear to be an event emitter
-							const needDefEmitter =
+							const needUse$on =
 								watcherCtx === component &&
-								!Object.isFunction(watcherCtx['on']) &&
-								!Object.isFunction(watcherCtx['addListener']);
+								!('on' in watcherCtx) &&
+								!('addListener' in watcherCtx);
 
-							if (needDefEmitter) {
+							if (needUse$on) {
 								unsafe.$on(watchPath, handler);
 
 							} else {
-								const addListener = (watcherCtx: ComponentInterface & EventEmitterLike) =>
+								const addListener = (watcherCtx: ComponentInterface & EventEmitterLike) => {
 									$a.on(watcherCtx, watchPath, <AnyFunction>handler, eventParams, ...watchInfo.args ?? []);
+								};
 
 								if (Object.isPromise(watcherCtx)) {
-									$a.promise<ComponentInterface & EventEmitterLike>(Object.cast(watcherCtx), asyncParams)
-										.then(addListener)
-										.catch(stderr);
+									type PromiseType = ComponentInterface & EventEmitterLike;
+									$a.promise<PromiseType>(Object.cast(watcherCtx), asyncParams).then(addListener).catch(stderr);
 
 								} else {
 									addListener(watcherCtx);
 								}
 							}
 
-							return;
-						}
-
-						const propInfo = p.info ?? getPropertyInfo(watchPath, component);
-
-						if (canSkipWatching(propInfo, watchInfo)) {
 							return;
 						}
 
@@ -299,14 +291,8 @@ export function bindRemoteWatchers(component: ComponentInterface, params?: BindR
 						/* eslint-enable prefer-const */
 
 						const emitter: EventEmitterLikeP = (_, wrappedHandler) => {
-							handler = Object.cast(wrappedHandler);
-
-							$a.worker(() => {
-								if (link != null) {
-									$a.off(link);
-								}
-							}, asyncParams);
-
+							handler = wrappedHandler;
+							$a.worker(() => link != null && $a.off(link), asyncParams);
 							return () => unwatch?.();
 						};
 
@@ -316,36 +302,28 @@ export function bindRemoteWatchers(component: ComponentInterface, params?: BindR
 
 				} else {
 					if (customWatcher) {
-						// True if the component itself can listen to an event,
-						// because watcherCtx does not appear to be an event emitter
-						const needDefEmitter =
+						const needUse$on =
 							watcherCtx === component &&
-							!Object.isFunction(watcherCtx['on']) &&
-							!Object.isFunction(watcherCtx['addListener']);
+							!('on' in watcherCtx) &&
+							!('addListener' in watcherCtx);
 
-						if (needDefEmitter) {
+						if (needUse$on) {
 							unsafe.$on(watchPath, handler);
 
 						} else {
-							const addListener = (watcherCtx: ComponentInterface & EventEmitterLike) =>
+							const addListener = (watcherCtx: ComponentInterface & EventEmitterLike) => {
 								$a.on(watcherCtx, watchPath, handler, eventParams, ...watchInfo.args ?? []);
+							};
 
 							if (Object.isPromise(watcherCtx)) {
-								$a.promise<ComponentInterface & EventEmitterLike>(Object.cast(watcherCtx), asyncParams)
-									.then(addListener)
-									.catch(stderr);
+								type PromiseType = ComponentInterface & EventEmitterLike;
+								$a.promise<PromiseType>(Object.cast(watcherCtx), asyncParams).then(addListener).catch(stderr);
 
 							} else {
 								addListener(watcherCtx);
 							}
 						}
 
-						return;
-					}
-
-					const propInfo = p.info ?? getPropertyInfo(watchPath, component);
-
-					if (canSkipWatching(propInfo, watchInfo)) {
 						return;
 					}
 
@@ -359,20 +337,26 @@ export function bindRemoteWatchers(component: ComponentInterface, params?: BindR
 
 					const emitter: EventEmitterLikeP = (_, wrappedHandler) => {
 						handler = Object.cast(wrappedHandler);
-
-						$a.worker(() => {
-							if (link != null) {
-								$a.off(link);
-							}
-						}, asyncParams);
-
+						$a.worker(() => link != null && $a.off(link), asyncParams);
 						return () => unwatch?.();
 					};
 
 					link = $a.on(emitter, 'mutation', handler, wrapWithSuspending(asyncParams, 'watchers'));
 					unwatch = $watch.call(component, propInfo, watchInfo, handler);
 				}
-			});
+			}
+		};
+
+		// Add a listener to the component's created hook if the component has not been created yet
+		if (attachWatcherOnCreated && isBeforeCreate) {
+			hooks['before:created'].push({fn: attachWatcher});
+
+		// Add a listener to the component's mounted/activated hook if the component has not been mounted or activated yet
+		} else if (attachWatcherOnMounted && (isBeforeCreate || component.$el == null)) {
+			hooks[isDeactivated ? 'activated' : 'mounted'].unshift({fn: attachWatcher});
+
+		} else {
+			attachWatcher();
 		}
-	});
+	}
 }